verl071/requirements.txt
# Core training framework
verl==0.7.1
torch==2.9.0+cu128
vllm==0.15.1
flash-attn==2.8.3

# ALFWorld environment
alfworld==0.4.2

# Supporting packages
transformers>=4.57.0
ray>=2.46.0
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0
